% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA792_climate_data.R
\docType{data}
\name{CA792_climate_data}
\alias{CA792_climate_data}
\title{Example Dataset: CA792_climate_data}
\format{
A SoilProfileCollection
}
\source{
CA792 Soil Survey
}
\usage{
data(CA792_climate_data)
}
\description{
A sample dataset containing climate data from the CA792 Soil Survey - Sequoia and Kings Canyon National Parks, created using
ecositerSpatial::site_prism_annual_normals()
}
\examples{
# Load the dataset
data(CA792_climate_data)

# Display the first few rows
head(CA792_climate_data)
}
\keyword{datasets}
