% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA792_veg_data.R
\docType{data}
\name{CA792_veg_data}
\alias{CA792_veg_data}
\title{Example Dataset: CA792_veg_data}
\format{
dataframe
}
\source{
CA792 Soil Survey
}
\usage{
data(CA792_veg_data)
}
\description{
A sample dataset containing veg data from the CA792 Soil Survey - Sequoia and Kings Canyon National Parks.
}
\examples{
# Load the dataset
data(CA792_veg_data)

# Display the first few rows
head(CA792_veg_data)
}
\keyword{datasets}
